# backend/openapi.yaml
openapi: 3.1.0
info:
  title: Graph Platform API
  version: 3.2.0
  description: API Graph Platform with Graph, FSM, Review, Ontology, RAG, Execution and Copilot engines.
servers:
  - url: http://localhost:3001/api
security:
  - bearerAuth: []
paths:
...
  /admin/users/{id}:
    patch: { operationId: adminUpdateUser, parameters: [{ $ref: '#/components/parameters/Id' }], requestBody: { $ref: '#/components/requestBodies/JsonBody' }, responses: { '200': { $ref: '#/components/responses/Success' } } }
    delete: { operationId: adminDeleteUser, parameters: [{ $ref: '#/components/parameters/Id' }], responses: { '200': { $ref: '#/components/responses/Success' } } }
  /admin/import-graph:
    post: { operationId: adminImportGraph, requestBody: { $ref: '#/components/requestBodies/JsonBody' }, responses: { '201': { $ref: '#/components/responses/Success' } } }
  /admin/nodes:
    post: { operationId: adminCreateNode, requestBody: { $ref: '#/components/requestBodies/JsonBody' }, responses: { '201': { $ref: '#/components/responses/Success' } } }
  /admin/nodes/{id}:
    delete: { operationId: adminDeleteNode, parameters: [{ $ref: '#/components/parameters/Id' }], responses: { '200': { $ref: '#/components/responses/Success' } } }
  /admin/edges:
    post: { operationId: adminCreateEdge, requestBody: { $ref: '#/components/requestBodies/JsonBody' }, responses: { '201': { $ref: '#/components/responses/Success' } } }
  /admin/edges/{id}:
    delete: { operationId: adminDeleteEdge, parameters: [{ $ref: '#/components/parameters/Id' }], responses: { '200': { $ref: '#/components/responses/Success' } } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string }
  requestBodies:
    JsonBody:
      required: true
      content:
        application/json:
          schema: { type: object, additionalProperties: true }
  responses:
    Success:
      description: Successful response
      content:
        application/json:
          schema: { type: [object, array], additionalProperties: true }